common: Only run mknod when root is running the script
mknod is needed by the vtpm test cases to be able to use dmesg via
/dev/kmsg. vtpm test cases require root, so nothing changes in terms
of requiring root now.

+++ common.sh
@@ -184,8 +184,10 @@ function setup_busybox_container()
   rootfs="$(get_busybox_container_root)"
 
   mkdir -p "${rootfs}"/{bin,mnt,proc,dev}
-  rm -f "${rootfs}"/dev/kmsg
-  mknod "${rootfs}"/dev/kmsg c 1 11
+  if [ "$(id -u)" = "0" ]; then
+    rm -f "${rootfs}"/dev/kmsg
+    mknod "${rootfs}"/dev/kmsg c 1 11
+  fi
 
   while [ $# -ne 0 ]; do
     if ! cp "$1" "${rootfs}"; then
